[ROCm] use the same fused_attn.h as NVTE, fixed a race runner issue in philox seed extraction, and add support matrix for aotriton
