+++ transformer_engine/jax/csrc/utils.h
@@ -15,11 +15,7 @@
 #include <string>
 #include <type_traits>
 #include <pybind11/pybind11.h>
-#ifndef USE_ROCM
 #include "transformer_engine/fused_attn.h"
-#else
-#include "transformer_engine/fused_attn_aotriton.h"
-#endif
 #include "common/util/logging.h"
 
 namespace transformer_engine {
